+++ lib/glimmer/swt/tab_item_proxy.rb
@@ -54,6 +54,12 @@ module Glimmer
           super(attribute_name)
         end
       end
+      
+      def dispose
+        swt_tab_item.setControl(nil)
+        swt_widget.dispose
+        swt_tab_item.dispose
+      end
     end
   end
 end